requirements.txt
alembic
azure-common
azure-nspkg
azure-storage-blob
azure-identity
azure-keyvault-secrets
certifi
cffi
chardet
Click
cryptography
Flask
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
idna
itsdangerous
Jinja2
Mako
MarkupSafe
msal
pycparser
pyodbc
pyopenssl
python-dateutil
python-editor
requests
setuptools
six
SQLAlchemy
urllib3
Werkzeug==2.2.2
WTForms
